+++ src/filters/cors.rs
@@ -20,7 +20,7 @@ use crate::reply::Reply;
 
 use self::internal::{CorsFilter, IntoOrigin, Seconds};
 
-/// Create a wrapping filter that exposes [CORS][] behavior for a wrapped
+/// Create a wrapping [`Filter`](crate::Filter) that exposes [CORS][] behavior for a wrapped
 /// filter.
 ///
 /// [CORS]: https://developer.mozilla.org/en-US/docs/Web/HTTP/CORS
@@ -56,7 +56,7 @@ pub fn cors() -> Builder {
     }
 }
 
-/// A wrapping filter constructed via `warp::cors()`.
+/// A wrapping [`Filter`](crate::Filter) constructed via `warp::cors()`.
 #[derive(Clone, Debug)]
 pub struct Cors {
     config: Arc<Configured>,